etl_pipeline/transfer_func/requirements.txt
functions-framework==3.4.0
google-cloud-storage==2.9.0
